[Fix] ModSpecs are now also set to something sensible for formats that are not editable by OpenMPT (useful when using e.g. the pattern vis in xmp-openmpt, since it relies on effect letters from ModSpecs) [Fix] DMO plugins could potentially crash in 64-bit builds.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@2938 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Vstplug.cpp
@@ -3031,7 +3031,7 @@ CDmo2Vst::CDmo2Vst(IMediaObject *pMO, IMediaObjectInPlace *pMOIP, DWORD uid)
 		m_Effect.numParams = dwParamCount;
 	}
 	if (FAILED(m_pMediaObject->QueryInterface(IID_IMediaParams, (void **)&m_pMediaParams))) m_pMediaParams = nullptr;
-	m_pMixBuffer = (int16 *)((((int)m_MixBuffer) + 15) & ~15);
+	m_pMixBuffer = (int16 *)((((intptr_t)m_MixBuffer) + 15) & ~15);
 	// Callbacks
 	m_Effect.dispatcher = DmoDispatcher;
 	m_Effect.setParameter = DmoSetParameter;
